requirements.txt
# Django Core
Django==5.0.7
asgiref==3.10.0
sqlparse==0.5.3
tzdata==2025.2


# Django REST Framework
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1

# Database
psycopg2-binary==2.9.11
dj-database-url==3.0.1

# CORS
django-cors-headers==4.3.1

# Authentication & Security
cryptography==46.0.3

# Payments
stripe==13.1.2

# Environment
python-dotenv==1.0.1

# Static Files
whitenoise==6.11.0

# Production Server
gunicorn==23.0.0

# Image Processing
pillow==10.4.0

# HTTP Requests (dependency for stripe and other packages)
requests==2.32.5
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
